Add files via upload
for right positioned tooltip:
 the tooltip is placed to the right (left:105%) of the "hoverable" text (<div>).
top:-5px is used to place it in the middle of its container element. We use the number 5 because the tooltip text has a top and bottom padding of 5px. If you increase its padding, also increase the value of the top property to ensure that it stays in the middle (if this is something you want). The same applies if you want the tooltip placed to the left.